typescript(vx-shape): consolidate and move stack and group types into types file
