+++ packages/vx-shape/src/shapes/Stack.tsx
@@ -13,9 +13,9 @@ import {
 
 import stackOrder, { STACK_ORDERS } from '../util/stackOrder';
 import stackOffset, { STACK_OFFSETS } from '../util/stackOffset';
+import { StackKey } from '../types';
 
 type Accessor<Datum> = (datum: Datum, index: number, data: Datum[]) => number;
-export type Key = string | number;
 
 export type StackProps<Datum> = {
   /** Array of data for which to generate a stack. */
@@ -29,14 +29,14 @@ export type StackProps<Datum> = {
   /** Sets the curve factory (from @vx/curve or d3-curve) for the area generator. Defaults to curveLinear. */
   curve?: CurveFactory;
   /** Returns a color for a given stack key and index. */
-  color?: (key: Key, index: number) => string;
+  color?: (key: StackKey, index: number) => string;
   /** Array of keys corresponding to stack layers. */
-  keys?: Key[];
+  keys?: StackKey[];
   /** Override render function which is passed the configured arc generator as input. */
   children?: (args: {
-    stacks: Series<Datum, Key>[];
+    stacks: Series<Datum, StackKey>[];
     path: AreaType<SeriesPoint<Datum>>;
-    stack: StackType<any, Datum, Key>;
+    stack: StackType<any, Datum, StackKey>;
   }) => React.ReactNode;
   /** Sets the x0 accessor function, and sets x1 to null. */
   x?: Accessor<SeriesPoint<Datum>>;
@@ -49,7 +49,7 @@ export type StackProps<Datum> = {
   /** Specifies the y1 accessor function which defaults to d => d[1]. */
   y1?: Accessor<SeriesPoint<Datum>>;
   /** Sets the value accessor for a Datum, which defaults to d[key]. */
-  value?: (d: Datum, key: Key) => number;
+  value?: (d: Datum, key: StackKey) => number;
   /** The defined accessor for the shape. The final area shape includes all points for which this function returns true. By default all points are defined. */
   defined?: (datum: SeriesPoint<Datum>, index: number, data: SeriesPoint<Datum>[]) => boolean;
   /** Sets the stack order to the pre-defined d3 function, see https://github.com/d3/d3-shape#stack_order. */
@@ -78,7 +78,7 @@ export default function Stack<Datum>({
   children,
   ...restProps
 }: StackProps<Datum> & Omit<React.SVGProps<SVGPathElement>, keyof StackProps<Datum>>) {
-  const stack = d3stack<Datum, Key>();
+  const stack = d3stack<Datum, StackKey>();
   if (keys) stack.keys(keys);
   if (value) stack.value(value);
   if (order) stack.order(stackOrder(order));
