+++ packages/vx-shape/src/types.ts
@@ -1,3 +1,74 @@
+import { SeriesPoint } from 'd3-shape';
+
+/** Unique key for item in a stack. */
+export type StackKey = string | number;
+
+/** Unique key for item in a group. */
+export type GroupKey = string | number;
+
+/** One BarGroup is returned for each datum, which has multiple sub-bars (based on keys). */
+export interface BarGroup {
+  /** index of BarGroup (matches input Datum index). */
+  index: number;
+  /** x0 position of bar group */
+  x0: number;
+  /** bars within group, one for each key. */
+  bars: BarGroupBar[];
+}
+
+/** One BarGroup is returned for each datum, which has multiple sub-bars (based on keys). */
+export interface BarGroupHorizontal {
+  /** index of BarGroup (matches input Datum index). */
+  index: number;
+  /** y0 position of bar group */
+  y0: number;
+  /** bars within group, one for each key. */
+  bars: BarGroupBar[];
+}
+
+export interface BarGroupBar {
+  /** group key */
+  key: GroupKey;
+  /** index of BarGroup (matches input Datum index). */
+  index: number;
+  /** group value (Datum[key]) */
+  value: number;
+  /** height of bar. */
+  height: number;
+  /** width of bar. */
+  width: number;
+  /** x position of bar. */
+  x: number;
+  /** y position of bar. */
+  y: number;
+  /** color of bar. */
+  color: string;
+}
+
+/** One BarStack is returned for each datum, which has multiple sub-bars (based on keys). */
+export interface BarStack<Datum> {
+  index: number;
+  key: StackKey;
+  bars: ({
+    /** Processed bar Datum with bar bounds and original datum. */
+    bar: SeriesPoint<Datum>;
+    /** group key */
+    key: StackKey;
+    /** index of BarGroup (matches input Datum index). */
+    index: number;
+    /** height of bar. */
+    height: number;
+    /** width of bar. */
+    width: number;
+    /** x position of bar. */
+    x: number;
+    /** y position of bar. */
+    y: number;
+    /** color of bar. */
+    color: string;
+  })[];
+}
+
 export type AccessorProps<Link, Node> = {
   /** Given a node, returns its x coordinate. */
   x?: (node: Node) => number;
